+++ modules/tools/editor/src/main/kotlin/dev/turingcomplete/intellijdevelopertoolsplugin/tool/editor/EditorSourceText.kt
@@ -4,6 +4,7 @@ import com.intellij.codeInsight.editorActions.SelectWordUtil
 import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.util.TextRange
 import dev.turingcomplete.intellijdevelopertoolsplugin.common.EditorUtils.getSelectedText
+import dev.turingcomplete.intellijdevelopertoolsplugin.common.EditorUtils.getSelectedTexts
 
 object EditorSourceText {
   // -- Properties ---------------------------------------------------------- //
@@ -12,6 +13,11 @@ object EditorSourceText {
 
   fun Editor.getSelectedTextOrTextAtCaret(): Pair<String, TextRange>? =
     getSelectedText() ?: getTextAtCaret()
+
+  fun Editor.getSelectedTextsOrTextAtCaret(): List<Pair<String, TextRange>> {
+    val selectedTexts = getSelectedTexts()
+    return selectedTexts.ifEmpty { listOfNotNull(getTextAtCaret()) }
+  }
 
   // -- Private Methods ----------------------------------------------------- //